# backend/docs/apifox_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Life Canvas OS API",
    "description": "\n# Life Canvas OS 后端 API\n\n八维生命平衡系统后端接口，支持个人成长管理、日记记录、AI 洞察等功能。\n\n## 主要功能\n\n- **用户认证**: PIN 码设置、验证、修改\n- **系统管理**: 八维系统（饮食、运动、智力、输出、恢复、资产、连接、环境）\n- **用户配置**: 用户信息、设置、AI 配置\n- **日记管理**: 日记 CRUD、附件、编辑历史\n- **AI 洞察**: 基于用户数据生成个性化洞察\n- **数据管理**: 备份、恢复、导出、导入\n\n## 认证方式\n\n本 API 使用 PIN 码认证。\n\n## 响应格式\n\n所有接口遵循统一的响应格式：\n\n```json\n{\n  \"code\": 200,\n  \"message\": \"success\",\n  \"data\": {},\n  \"timestamp\": 1707219200000\n}\n```\n\n## 错误码\n\n- `200`: 成功\n- `201`: 创建成功\n- `400`: 错误请求\n- `401`: 未授权\n- `404`: 资源不存在\n- `409`: 冲突\n- `422`: 参数验证失败\n- `429`: 请求过于频繁\n- `500`: 服务器错误\n\n## 分页参数\n\n| 参数 | 类型 | 必填 | 默认值 | 说明 |\n|------|------|------|--------|------|\n| page | Integer | 否 | 1 | 页码，从 1 开始 |\n| page_size | Integer | 否 | 20 | 每页数量，最大 100 |\n| sort_by | String | 否 | created_at | 排序字段 |\n| sort_order | String | 否 | desc | 排序方向：asc/desc |\n",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/user/profile": {
      "get": {
        "tags": [
          "users",
          "user"
        ],
        "summary": "获取用户信息",
        "description": "获取用户信息",
        "operationId": "get_user_profile_api_user_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserResponse"
                    },
                    "timestamp": {
                      "type": "integer",
                      "example": 1707219200000
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "users",
          "user"
        ],
        "summary": "更新用户信息",
        "description": "更新用户信息",
        "operationId": "update_user_profile_api_user_profile_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserResponse"
                    },
                    "timestamp": {
                      "type": "integer",
                      "example": 1707219200000
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/settings": {
      "get": {
        "tags": [
          "users",
          "user"
        ],
        "summary": "获取用户设置",
        "description": "获取用户设置",
        "operationId": "get_user_settings_api_user_settings_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserSettingsResponse"
                    },
                    "timestamp": {
                      "type": "integer",
                      "example": 1707219200000
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "users",
          "user"
        ],
        "summary": "更新用户设置",
        "description": "更新用户设置",
        "operationId": "update_user_settings_api_user_settings_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserSettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserSettingsResponse"
                    },
                    "timestamp": {
                      "type": "integer",
                      "example": 1707219200000
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/user/ai-config": {
      "get": {
        "tags": [
          "users",
          "user"
        ],
        "summary": "获取AI配置",
        "description": "获取 AI 配置（不返回完整 API Key）",
        "operationId": "get_ai_config_api_user_ai_config_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "$ref": "#/components/schemas/AIConfigResponse"
                    },
                    "timestamp": {
                      "type": "integer",
                      "example": 1707219200000
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "users",
          "user"
        ],
        "summary": "保存AI配置",
        "description": "保存 AI 配置",
        "operationId": "save_ai_config_api_user_ai_config_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AIConfigBase"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "integer",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "success"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "provider": {
                          "type": "string"
                        },
                        "model_name": {
                          "type": "string"
                        },
                        "updated_at": {
                          "type": "string"
                        }
                      }
                    },
                    "timestamp": {
                      "type": "integer",
                      "example": 1707219200000
                    }
                  }
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AIConfigBase": {
        "properties": {
          "provider": {
            "type": "string",
            "enum": [
              "deepseek",
              "doubao",
              "openai"
            ],
            "title": "Provider"
          },
          "api_key": {
            "type": "string",
            "minLength": 1,
            "title": "Api Key",
            "description": "API Key"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name",
            "description": "模型名称"
          }
        },
        "type": "object",
        "required": [
          "provider",
          "api_key"
        ],
        "title": "AIConfigBase",
        "description": "AI 配置基础"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "UserSettingsUpdate": {
        "properties": {
          "theme": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "light",
                  "dark",
                  "auto"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Theme"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "auto_save_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Save Enabled"
          },
          "auto_save_interval": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 600.0,
                "minimum": 10.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Auto Save Interval"
          },
          "notification_enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notification Enabled"
          },
          "notification_time": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\d{2}:\\d{2}$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notification Time"
          },
          "show_year_progress": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Show Year Progress"
          },
          "show_weekday": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Show Weekday"
          }
        },
        "type": "object",
        "title": "UserSettingsUpdate",
        "description": "更新用户设置（所有字段可选）"
      },
      "UserUpdate": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "birthday": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Birthday",
            "description": "生日字符串 YYYY-MM-DD"
          },
          "mbti": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 4,
                "minLength": 4
              },
              {
                "type": "null"
              }
            ],
            "title": "Mbti"
          },
          "values": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Values"
          },
          "life_expectancy": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 120.0,
                "minimum": 50.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Life Expectancy"
          }
        },
        "type": "object",
        "title": "UserUpdate",
        "description": "更新用户信息（所有字段可选）"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "username": {
            "type": "string"
          },
          "display_name": {
            "type": "string"
          },
          "birthday": {
            "type": "string",
            "format": "date"
          },
          "mbti": {
            "type": "string"
          },
          "values": {
            "type": "string"
          },
          "life_expectancy": {
            "type": "integer"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "UserSettingsResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "user_id": {
            "type": "integer"
          },
          "theme": {
            "type": "string",
            "enum": [
              "light",
              "dark",
              "auto"
            ]
          },
          "language": {
            "type": "string"
          },
          "auto_save_enabled": {
            "type": "boolean"
          },
          "auto_save_interval": {
            "type": "integer"
          },
          "notification_enabled": {
            "type": "boolean"
          },
          "notification_time": {
            "type": "string"
          },
          "show_year_progress": {
            "type": "boolean"
          },
          "show_weekday": {
            "type": "boolean"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AIConfigResponse": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "enum": [
              "deepseek",
              "doubao",
              "openai"
            ]
          },
          "model_name": {
            "type": "string"
          },
          "api_key_masked": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000",
      "description": "开发环境"
    }
  ],
...
}